Structure change
Changed the structure of the code
Events have now being segrehated except "ready"
Code under "commands" have been aligned for better readability
Added dotenv for better dev experience
Shortened index.js
Added gitignore for git ignoring files
