+++ commands/beep.js
@@ -1,7 +1,7 @@
 module.exports = {
 	name: 'beep',
 	description: 'Beep!',
-  eligible: 1,
+	eligible: 1,
 	execute(message) {
 		message.channel.send('Boop.');
 	},
