+++ commands/help.js
@@ -1,4 +1,4 @@
-const prefix= process.env.PREFIX;
+const prefix = process.env.PREFIX;
 const Discord = require('discord.js');
 module.exports = {
 	name: 'help',
@@ -6,20 +6,20 @@ module.exports = {
 	aliases: ['commands'],
 	usage: '[command name]',
 	cooldown: 5,
-  eligible: 1,
+	eligible: 1,
 	execute(message, args) {
 		const data = [];
 		const { commands } = message.client;
 
 		if (!args.length) {
-let helpEmbed = new Discord.MessageEmbed()
-  .setColor(0x4286f4)
-  .setURL(process.env.URL)
-  .setTitle("List of all my commands")
-  .setDescription("`" + commands.map(command => command.name).join('`, `') + "`")
+			let helpEmbed = new Discord.MessageEmbed()
+				.setColor(0x4286f4)
+				.setURL(process.env.URL)
+				.setTitle("List of all my commands")
+				.setDescription("`" + commands.map(command => command.name).join('`, `') + "`")
 
-  .addField('Usage', `\nYou can send \`${prefix}help [command name]\` to get info on a specific command!`);
-return message.author.send(helpEmbed)
+				.addField('Usage', `\nYou can send \`${prefix}help [command name]\` to get info on a specific command!`);
+			return message.author.send(helpEmbed)
 
 				.then(() => {
 					if (message.channel.type === 'dm') return;
@@ -38,15 +38,15 @@ return message.author.send(helpEmbed)
 			return message.reply('that\'s not a valid command!');
 		}
 
-let commandEmbed = new Discord.MessageEmbed()
-.setColor(0x4286f4)
-.setTitle("Command Help")
+		let commandEmbed = new Discord.MessageEmbed()
+			.setColor(0x4286f4)
+			.setTitle("Command Help")
 
-if (command.description) commandEmbed.setDescription(`${command.description}`)
+		if (command.description) commandEmbed.setDescription(`${command.description}`)
 
-if (command.aliases) commandEmbed.addField('Aliases', `\`${command.aliases.join(', ')}\``, true)
-.addField('Cooldown', `${command.cooldown || 3} second(s)`, true);
-if (command.usage) commandEmbed.addField('Usage', `\`${prefix}${command.name} ${command.usage}\``, true)
+		if (command.aliases) commandEmbed.addField('Aliases', `\`${command.aliases.join(', ')}\``, true)
+			.addField('Cooldown', `${command.cooldown || 3} second(s)`, true);
+		if (command.usage) commandEmbed.addField('Usage', `\`${prefix}${command.name} ${command.usage}\``, true)
 		message.channel.send(commandEmbed);
 	},
 };
