+++ commands/server.js
@@ -1,7 +1,7 @@
 module.exports = {
 	name: 'server',
 	description: 'Display info about this server.',
-  eligible: 1,
+	eligible: 1,
 	execute(message) {
 		message.channel.send(`Server name: ${message.guild.name}\nTotal members: ${message.guild.memberCount}`);
 	},
