+++ commands/user-info.js
@@ -1,7 +1,7 @@
 module.exports = {
 	name: 'user-info',
 	description: 'Display info about yourself.',
-  eligible: 1,
+	eligible: 1,
 	execute(message) {
 		message.channel.send(`Your username: ${message.author.username}\nYour ID: ${message.author.id}`);
 	},
