+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "SMR-Bot",
+  "name": "smr-bot",
   "version": "0.2.0",
   "description": "A all purpose Discord Bot",
   "main": "index.js",
@@ -18,6 +18,7 @@
   "license": "ISC",
   "dependencies": {
     "discord.js": "^12.3.1",
+    "dotenv": "^8.2.0",
     "node-os-utils": "^1.3.2"
   }
 }
